+++ src/main/java/dta/sfmflow/block/ModBlocks.java
@@ -73,7 +73,7 @@ public class ModBlocks {
 	
 	public static final DeferredBlock<Block> SCULK_TRIGGER_CABLE_BLOCK = registerBlock("sculk_trigger_cable_block",
 			() -> new SculkTriggerCableBlock(BlockBehaviour.Properties.of().strength(2.0F, 5.0F).sound(SoundType.METAL)
-					.requiresCorrectToolForDrops().pushReaction(PushReaction.BLOCK)));
+					.requiresCorrectToolForDrops().pushReaction(PushReaction.BLOCK).noOcclusion()));
 
 	private static <T extends Block> DeferredBlock<T> registerBlock(String name, Supplier<T> block) {
 		DeferredBlock<T> toReturn = BLOCKS.register(name, block);