+++ src/main/java/dta/sfmflow/datagen/ModBlockLootTableProvider.java
@@ -32,6 +32,7 @@ public class ModBlockLootTableProvider extends BlockLootSubProvider {
 		dropSelf(ModBlocks.FLUID_VACUUM_VALVE_BLOCK.get());
 		dropSelf(ModBlocks.CABLE_CLUSTER_BLOCK.get());
 		dropSelf(ModBlocks.ADVANCED_CABLE_CLUSTER_BLOCK.get());
+		dropSelf(ModBlocks.SCULK_TRIGGER_CABLE_BLOCK.get());
 	}
 
 	@Override